requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.7
django-anymail==13.1
django-environ==0.12.0
django-jazzmin==3.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
requests==2.32.5
sqlparse==0.5.3
stripe==13.0.1
typing_extensions==4.15.0
urllib3==2.5.0
